Feedback and training modifications

+++ Services/TrainingRegistrationService.cs
@@ -18,7 +18,7 @@ public interface ITrainingRegistrationService
     TrainingRegistration GetById(int id, Boolean expand = false);
     IEnumerable<TrainingRegistrationViewMax> GetUserTrainingHistory(int userid);
     void Update(int id, UpdateRequest model);
-    void Delete(int id);
+    void Delete(int userId, string roleId, string teamId, int id);
 }
 
 public class TrainingRegistrationService : ITrainingRegistrationService
@@ -45,15 +45,17 @@ public class TrainingRegistrationService : ITrainingRegistrationService
     public IEnumerable<TrainingRegistrationView> GetAll(Boolean expand = false)
     {
         if (expand == true)
-            return _context.TrainingRegistration.Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Select(tr => new TrainingRegistrationViewMax {
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id,
-                Training=tr.Training,
-                Status=tr.Status,
-                User = new UserView{
+            return _context.TrainingRegistration.Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Select(tr => new TrainingRegistrationViewMax
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id,
+                Training = tr.Training,
+                Status = tr.Status,
+                User = new UserView
+                {
                     id = tr.User.id,
                     username = tr.User.username,
                     firstname = tr.User.firstname,
@@ -62,12 +64,13 @@ public class TrainingRegistrationService : ITrainingRegistrationService
                 }
             });
         else
-            return _context.TrainingRegistration.Select(tr => new TrainingRegistrationView{
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id
+            return _context.TrainingRegistration.Select(tr => new TrainingRegistrationView
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id
             });
     }
 
@@ -75,21 +78,26 @@ public class TrainingRegistrationService : ITrainingRegistrationService
     {
         if (teamId == null) throw new KeyNotFoundException("Invalid teamid.");
         Team team = null;
-        try {
+        try
+        {
             team = _sharedService.GetTeam(Byte.Parse(teamId));
-        } catch {
+        }
+        catch
+        {
             throw new KeyNotFoundException("Invalid teamid.");
         }
         if (expand == true)
-            return _context.TrainingRegistration.Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationViewMax {
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id,
-                Training=tr.Training,
-                Status=tr.Status,
-                User = new UserView{
+            return _context.TrainingRegistration.Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationViewMax
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id,
+                Training = tr.Training,
+                Status = tr.Status,
+                User = new UserView
+                {
                     id = tr.User.id,
                     username = tr.User.username,
                     firstname = tr.User.firstname,
@@ -98,27 +106,30 @@ public class TrainingRegistrationService : ITrainingRegistrationService
                 }
             });
         else
-            return _context.TrainingRegistration.Include(t => t.User).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationView{
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id
+            return _context.TrainingRegistration.Include(t => t.User).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationView
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id
             });
     }
 
     public IEnumerable<TrainingRegistrationView> GetAllByTraining(int trainingId, Boolean expand = false)
     {
         if (expand == true)
-            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Include(t => t.User.Team).Include(t => t.User.Role).Select(tr => new TrainingRegistrationViewMaxExpand {
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id,
-                Training=tr.Training,
-                Status=tr.Status,
-                User = new UserViewExpand{
+            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Include(t => t.User.Team).Include(t => t.User.Role).Select(tr => new TrainingRegistrationViewMaxExpand
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id,
+                Training = tr.Training,
+                Status = tr.Status,
+                User = new UserViewExpand
+                {
                     id = tr.User.id,
                     username = tr.User.username,
                     firstname = tr.User.firstname,
@@ -129,12 +140,13 @@ public class TrainingRegistrationService : ITrainingRegistrationService
                 }
             });
         else
-            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Select(tr => new TrainingRegistrationView{
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id
+            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Select(tr => new TrainingRegistrationView
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id
             });
     }
 
@@ -142,21 +154,26 @@ public class TrainingRegistrationService : ITrainingRegistrationService
     {
         if (teamId == null) throw new KeyNotFoundException("Invalid teamid.");
         Team team = null;
-        try {
+        try
+        {
             team = _sharedService.GetTeam(Byte.Parse(teamId));
-        } catch {
+        }
+        catch
+        {
             throw new KeyNotFoundException("Invalid teamid.");
         }
         if (expand == true)
-            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Include(t => t.User.Team).Include(t => t.User.Role).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationViewMaxExpand {
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id,
-                Training=tr.Training,
-                Status=tr.Status,
-                User = new UserViewExpand{
+            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Include(t => t.User.Team).Include(t => t.User.Role).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationViewMaxExpand
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id,
+                Training = tr.Training,
+                Status = tr.Status,
+                User = new UserViewExpand
+                {
                     id = tr.User.id,
                     username = tr.User.username,
                     firstname = tr.User.firstname,
@@ -167,12 +184,13 @@ public class TrainingRegistrationService : ITrainingRegistrationService
                 }
             });
         else
-            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Include(t => t.User).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationView{
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id
+            return _context.TrainingRegistration.Where(u => u.training_id == trainingId).Include(t => t.User).Where(u => u.User.team_id == team.id).Select(tr => new TrainingRegistrationView
+            {
+                id = tr.id,
+                user_id = tr.user_id,
+                training_id = tr.training_id,
+                registration_date = tr.registration_date,
+                status_id = tr.status_id
             });
     }
 
@@ -182,7 +200,8 @@ public class TrainingRegistrationService : ITrainingRegistrationService
             throw new AppException("User does not exist");
         if (!_context.Training.Any(x => x.id == model.training_id))
             throw new AppException("Training does not exist");
-        if (_context.TrainingRegistration.Any(x => x.training_id == model.training_id && x.user_id == model.user_id)) {
+        if (_context.TrainingRegistration.Any(x => x.training_id == model.training_id && x.user_id == model.user_id))
+        {
             throw new AppException("User already registered to this training");
         }
         var trainingRegistration = _mapper.Map<TrainingRegistration>(model);
@@ -192,22 +211,25 @@ public class TrainingRegistrationService : ITrainingRegistrationService
         _context.SaveChanges();
     }
 
-    public IEnumerable<TrainingRegistrationViewMax> GetUserTrainingHistory(int userid) {
-        return _context.TrainingRegistration.Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Where(t => t.user_id == userid).Select(tr => new TrainingRegistrationViewMax {
-                id=tr.id,
-                user_id=tr.user_id,
-                training_id=tr.training_id,
-                registration_date=tr.registration_date,
-                status_id=tr.status_id,
-                Training=tr.Training,
-                Status=tr.Status,
-                User = new UserView{
-                    id = tr.User.id,
-                    username = tr.User.username,
-                    firstname = tr.User.firstname,
-                    lastname = tr.User.lastname,
-                    email = tr.User.email,
-                }
+    public IEnumerable<TrainingRegistrationViewMax> GetUserTrainingHistory(int userid)
+    {
+        return _context.TrainingRegistration.Include(t => t.Status).Include(t => t.Training).Include(t => t.User).Where(t => t.user_id == userid).Select(tr => new TrainingRegistrationViewMax
+        {
+            id = tr.id,
+            user_id = tr.user_id,
+            training_id = tr.training_id,
+            registration_date = tr.registration_date,
+            status_id = tr.status_id,
+            Training = tr.Training,
+            Status = tr.Status,
+            User = new UserView
+            {
+                id = tr.User.id,
+                username = tr.User.username,
+                firstname = tr.User.firstname,
+                lastname = tr.User.lastname,
+                email = tr.User.email,
+            }
         });
     }
 
@@ -227,10 +249,39 @@ public class TrainingRegistrationService : ITrainingRegistrationService
     }
 
 
-    public void Delete(int id)
+    public void Delete(int userId, string roleId, string teamId, int id)
     {
-        var trainingRegistration = _sharedService.GetTrainingRegistration(id);
-        _context.TrainingRegistration.Remove(trainingRegistration);
+        TrainingRegistration trainingRegistration = null;
+        switch (roleId)
+        {
+            case "1":
+                trainingRegistration = _context.TrainingRegistration.Find(id);
+                if (trainingRegistration == null) throw new KeyNotFoundException("Training registration not found");
+                if (trainingRegistration.user_id != userId) throw new AppException("You do not have access to this registration");
+                _context.TrainingRegistration.Remove(trainingRegistration);
+                break;
+            case "2":
+                if (teamId == null) throw new KeyNotFoundException("Invalid teamid.");
+                Team team = null;
+                try
+                {
+                    team = _sharedService.GetTeam(Byte.Parse(teamId));
+                }
+                catch
+                {
+                    throw new KeyNotFoundException("Invalid teamid.");
+                }
+                trainingRegistration = _context.TrainingRegistration.Include(tr => tr.User).Where(tr => tr.id == id).FirstOrDefault();
+                if (trainingRegistration == null) throw new KeyNotFoundException("Training registration not found");
+                if (trainingRegistration.User.team_id != team.id) throw new AppException("You do not have access to this registration");
+                _context.TrainingRegistration.Remove(trainingRegistration);
+                break;
+            case "3":
+                trainingRegistration = _context.TrainingRegistration.Find(id);
+                if (trainingRegistration == null) throw new KeyNotFoundException("Training registration not found");
+                _context.TrainingRegistration.Remove(trainingRegistration);
+                break;
+        }
         _context.SaveChanges();
     }
